Address feedback to simplify styles
